+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 
+* Add `--aggregate` flag for grouping all rewards by day.
 * Bump dependency versions(text).